requirements.txt
# IP Address Service - Python Dependencies
# Core web framework
Flask==3.0.0
Werkzeug==3.0.1

# Production WSGI server
gunicorn==21.2.0

# Additional utilities for production deployment
python-dotenv==1.0.0

# Security and validation
ipaddress==1.0.23  # Built-in but explicit for documentation

# Development and testing dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

